noodle: introduce sourcegraph literal search mentions provider
Note: this logic is currently hidden behind the noodle feature flag.
Additionally instead of implementing client logic in our graphql
backend, it keeps all related logic inside one file to contain the blast
radius.

When typing the prefix "@src:" we will provide a context item which says
it will search against your sourcegraph endpoint. This is presented as a
URL to the search. When clicking resolving the item it will convert each
search result chunk into a context item.

This is currently not working due to some recent restrictions introduced
on context sources. Note that the URL mention provider also seems to be
broken due to it:

  Request Failed: The prompt string contains a reference to a file that
  is not allowed by the context filters.

Other known issues:
- We can't use the triggerPrefix ! or ? due to how mentions are parsed
- Mentions also stop completing once you enter non file path looking
  items. EG entering a ( stops the autocomplete.
- The generated link to the search results is not clickable due to
  vscode incorrectly escaping the query parameters.

Test Plan: manually tested, but not yet ready for merging