Onboarding tooltip: don't hide the code (#1369)
Previously, there were cases when the position of the onboarding tooltip
would hide the underlying code. This goes against JB recommendations
https://jetbrains.design/intellij/controls/got_it_tooltip/#location even
if we are using the built-in position helpers. This PR changes the
position dynamically based on

- Whether the project window is open
- Whether the cursor is close to the top

Depending on the different values, we now either display the tooltip to
the left, above, or below.

## Test plan

Manually tested

Project view closed and at the top of the file

![CleanShot 2024-04-23 at 11 26
57@2x](https://github.com/sourcegraph/jetbrains/assets/1408093/484bedbf-1789-4f19-90b7-b36fe52970b6)

Project view is open at the top

![CleanShot 2024-04-23 at 11 40
19@2x](https://github.com/sourcegraph/jetbrains/assets/1408093/b1627383-821c-4a7c-8806-de1ce75fbd25)

Project view is closed but there is space above

![CleanShot 2024-04-23 at 11 40
55@2x](https://github.com/sourcegraph/jetbrains/assets/1408093/f2f87c29-3c53-49ef-bda4-756490ad79c8)



<!-- All pull requests REQUIRE a test plan:
https://sourcegraph.com/docs/dev/background-information/testing_principles

Why does it matter?

These test plans are there to demonstrate that are following industry
standards which are important or critical for our customers.
They might be read by customers or an auditor. There are meant be simple
and easy to read. Simply explain what you did to ensure
your changes are correct!

Here are a non exhaustive list of test plan examples to help you:

- Making changes on a given feature or component:
- "Covered by existing tests" or "CI" for the shortest possible plan if
there is zero ambiguity
  - "Added new tests"
- "Manually tested" (if non trivial, share some output, logs, or
screenshot)
- Updating docs:
  - "previewed locally"
  - share a screenshot if you want to be thorough
- Updating deps, that would typically fail immediately in CI if
incorrect
  - "CI"
  - "locally tested"
-->