Try alternative UX (style button)
The # is potentially a bit difficult to discover, so trying a simple UI element instead to tweak response style. The idea is still the same though where you pick a bunch of style-tags to apply.

This can later always be expanded to ALSO allow for quick access via a #-like interface so you can simply type everything instead of having to click