Add a service for checking Cody Ignore policy and decorate some UI with ignored status (#1379)
This change is integrated through Agent, but the extension side is not
integrated with server-set policy yet. Instead, this change uses a
testing hook which pushes an ad-hoc policy to the extension—see the test
plan.

As a result, this does not cause any user visible change for Enterprise
users yet. For self-serve users of the experimental file-based Cody
Ignore feature, this UX may trigger until extension-side integration is
complete. However we have determined that to be very few users.

Exact UX is still being designed. These labels and icons are
placeholders.

Part of #1252, #1253.

## Test plan

Tested manually:

1. Run the extension with
`CODY_JETBRAINS_FEATURES=cody.feature.internals-menu=true` and open some
files, say `Foo.java` and `Bar.go`
2. Status bar, Internals, Testing: Cody Ignore
3. Check "Override policy for testing", set URI regex to `\.java$`. OK.
4. Verify that a banner appears in `Foo.java` but not `Bar.go`.
5. Run a command like Explain Code with the right click context menu and
verify that a notification appears
6. Run a command like Explain Code with a keyboard shortcut like
ctrl-shift-E and verify that a notification appears
7. Verify when focus switches from `Foo.java` to `Bar.go`, a slashed
Cody icon appears in the status bar when an ignored file is focused.