Overhaul of the Edit Code dialog, hotkeys, and Edit/Retry (#1372)
This PR introduces a rewrite of the **Edit Code with Cody** dialog to
match Daniel's designs. The dialog has changed into a JFrame to give it
maximum flexibility and customizability, though at the cost of having to
write a lot of code to do things you get for free in dialogs.

Overview of the PR:

- All the theming has been redone to match the designs
- The input text field is now multi-line and scrollable
- The dialog is movable and resizable (although resizing is currently
janky)
- Dialog handles theme changes correctly (though now you can't see
because it closes on loss of focus)
- Dialog opens at the line where you invoked it
- All code lenses are registered as actions and have working keyboard
shortcuts
- All keyboard shortcuts are rendered properly rather than being
hardcoded
- The dialog's history handling has been cleaned up

I've made a start at re-theming the code lenses, but it has another day
or two of work. I'll add it to this PR.

## Test plan

I created a test plan for the instructions dialog in `TESTING.md`. It is
pretty comprehensive and has dozens of tests and assertions.

We _really_ need UI/Robot tests for the UI, which is already starting to
become complicated.