Cody Ignore: Send "ignore" capability, update hook for new test override syntax (#1396)
Fixes #1393

Part of #1252 #1253

Depends on sourcegraph/cody#3946

## Test plan

Tested manually:

1. Run the extension with
`CODY_JETBRAINS_FEATURES=cody.feature.internals-menu=true` and open the
sourcegraph/sourcegraph repo.
2. Status bar, Internals, Testing: Cody Ignore
3. Check "Override policy for testing". OK.
4. Verify that no banners appear
5. Change the policy from excluding `...sourcegraph/cody` to
`...sourcegraph/sourcegraph`
6. Verify that banners appear on open editors
7. Run a command like Explain Code with the right click context menu and
verify that a notification appears
8. Run a command like Explain Code with a keyboard shortcut like
ctrl-shift-E and verify that a notification appears
9. Verify when focus switches from a file within sourcegraph/sourcegraph
to a file outside it, a slashed Cody icon appears in the status bar when
an ignored file is focused.
10. Disable the policy override, verify that the banners/blocked
commands/etc. revert.