noodle: include patternType in search URL
This is so when you click through it uses the same interpretation of the
query string we use against the backend.

Additionally instead of using the title of the item, we extract out the
query from the URI. This is a small change which makes this more useful
as a generic extractor of search results from URLs.