noodle: Explain Code History command
This is a super experimental command which relies on the output of "git
log". It uses git logs support to understand function boundaries such
that we can give commit history only for commits which change a
function's body.

I spent a little while working out how to use things like the git
extension API/etc, but this doesn't really work along those lines.
However, this sort of functionality is implementable in gitserver so
this shouldn't limit us there.

Test Plan: right clicked on a few functions and eye balled the output.