TSC Graph Context: handle property access
Previously, the `tsc-mixed` context didn't boost the type of the
*expression* (aka. qualifier) in property accesses. For example, the
property access
```ts
this.helper().CURSOR
```
has the *expression* `this.helper()`, which might have the type `interface
Helper { ... }` with relevant members for autocomplete. This PR fixes
the problem by effectively doing "Go to type definition" on the
expression and including that definition in the context.