TSC Graph Context: remove `typescript` from bundle
Fixes CODY-1294

Previously, the bundle size of the VSC extension was ~21mb instead of
~14mb because of the embedded typescript dependency. This was
problematic because parsing the JS bundle was slow and could cause the
editor to occasionally freeze. This PR removes the `typescript`
dependency via the `--external:typescript` esbuild flag. Surprisingly,
the `tsc-mixed` context strategy still seems to work within VSC and it
might be because VSC somehow exposes access to TypeScript APIs by
default similar to how it works for `vscode` imports. However, this
change means that `tsc-mixed` does not work for the agent until we
bundle `typescript.js` alongside the `index.js` for the agent.