noodle: add ContextItemHistory type
This is built on top of the explain history PR to introduce a type for
history. Right now it is hacked in as pretend terminal output and
doesn't work that well. This is the start of presenting the context more
clearly.

This is still early, so is a draft.

Test Plan: CI