Autocomplete: force 17b on explicit autocomplete with starcoder-hybrid
Previously, when using starcoder-hybrid, we would use starcoder-7b for
single-line completions even when the user explicitly triggers
autocomplete via shortcut. This PR changes the behavior so we force
starcoder-16b on explicit trigger.

The reasoning behind this change is that users are OK with getting
longer latency when they explicitly trigger the model.

A possible followup change is to enable multi-line completion.