Finished up the Edit/Retry feature (#1439)
Edit/Retry had a couple of issues, both fixed:

1. It wasn't Undoing before retrying, so it was leaving the original
text in the document. Pretty bad bug.
2. Document Code needed to show the "instruction" (prompt) generated by
the Agent. This mostly a snazzy extra from VS Code.

This PR goes hand-in-hand with a [PR in
sourcegraph/cody](https://github.com/sourcegraph/cody/pull/4010) that
sends the instruction along with task updates, so that it's available if
the user chooses to retry.

This PR also fixes a couple of bugs, including a really annoying one
with Undo, where it was scrolling to the top of the document.

## Test plan

Test plan is in TESTING.md.