Fix listeners calls to agent (#1440)
## Changes 

When reading agent code I realised that change in cursor/selection
should in fact be passed by `didChange` and not `didFocus` event. This
was changed recently by me while trying to improve situation with agent
crashes (and when we though excessive amount of agent notifications
might be a partial reason) but I realised this change was simply
incorrect.


## Test plan

Manual check with debugger if proper endpoints are hit.