swallow hidePopover exception
Attempt to fix issue reported by 2 people where the Cody webview was blank in VS Code and the following error appeared in the devtools console:

```
Uncaught DOMException: Failed to execute 'hidePopover' on 'HTMLElement': Invalid on popover elements that aren't already showing. This might have been the result of the "beforetoggle" event handler changing the state of this popover.
```