allow scrolling with mouse of chat message editor
Previously, when the user had entered a very long message in chat (with many lines, specifically more than 80% of the viewport height), it was not possible to scroll the chat message editor up/down with the mouse or trackpad. This was a bug. (Arrow keys would still scroll it by moving the text insertion point.)

Now, it is possible to scroll with the mouse or trackpad.

Reported by @martinamps