slightly clearer message in context popover for git repo subdirs
If you open a subdir of a Git repository as the VS Code workspace root folder, then Cody will show `Folder is not a Git repository` as the embeddings status in the enhanced context popover. This is true, but it can be confusing. The ideal behavior is to make embeddings "just work" in this case, but in the meantime, this message is slightly clearer and should indicate to users that embeddings would work if they open up the Git repository root dir in the editor.

Reported by @martinamps