Edit: introduce CodyTaskState.Reverting and FixupTask.retryID
I am introducing an automation on top of Fixup and need to be able to
track when a Fixup for a region is done. "retry" is actually an "undo"
on the old FixupTask and then a new FixupTask is created. Before this
commit there was no way to link the old task to the new retried task.

This commit introduces a field on FixupTask which records the new task's
ID. Additionally a "Reverting" state is introduced which occurs while
undoing a task before it enters a terminal state. We then adjust "retry"
to only set the terminal state once the new FixupTask is created. This
makes it possible to set the retryID field before updating the task a
terminal state.

Test Plan: Added the following debug line to FixupController.createTask

  task.onDidStateChange(state => {
      console.log('onDidStateChange', task.id, state, task.retryID)
  })

First did an edit followed by an undo. Expectation is that for the undo
we now have a Reverting state introduced:

  onDidStateChange lwszk Working undefined
  onDidStateChange lwszk Applying undefined
  onDidStateChange lwszk Formatting undefined
  onDidStateChange lwszk Applied undefined
  onDidStateChange lwszk Reverting undefined
  onDidStateChange lwszk Finished undefined

Secondly did a edit followed by a retry. Expectation is we enter
reverting state, create the new task, finish the old task with retryID
set, then continue with the new task:

  onDidStateChange lwszqu Working undefined
  onDidStateChange lwszqu Applying undefined
  onDidStateChange lwszqu Formatting undefined
  onDidStateChange lwszqu Applied undefined
  onDidStateChange lwszqu Reverting undefined
  onDidStateChange lwtj Working undefined
  onDidStateChange lwtj Applying undefined
  onDidStateChange lwszqu Finished lwtj
  onDidStateChange lwtj Formatting undefined
  onDidStateChange lwtj Applied undefined