use simpler way of loading OpenCtx (for mentions/context)
Previously, Cody would try to communicate with the OpenCtx VS Code extension in the extension host process. This required both the Cody and OpenCtx VS Code extensions to both be installed and activated, which was tricky and made it hard to develop on both simultaneously.

Now, Cody uses [@openctx/vscode-lib](https://www.npmjs.com/package/@openctx/vscode-lib), which means it does not need to rely on the OpenCtx VS Code extension being installed and activated.

To avoid affecting users who are not participating in this experiment, you must set `cody.experimental.openctx` to true in your VS Code user settings to use OpenCtx in Cody.

To try it out end-to-end, add the following to your VS Code user settings:

```
{
  "openctx.enable": true,
  "openctx.debug": true,
  "openctx.providers": {
    "https://openctx.org/npm/@openctx/provider-hello-world": true,
  },
  "cody.debug.verbose": true,
  "cody.experimental.openctx": true
  "cody.experimental.noodle": true,
}
```

And then in a new chat window, type `@`, select `OpenCtx` from the mention menu, and look for the `Hello, world!` mention item. This will currently only work in a VS Code debug extension host.