fix bug where `Add Selection to Cody Chat` would sometimes not work
If the chat message editor had no selection, which is a possible state for it to get into, then the insertion would fail. Instead, we can just use Lexical's `$insertNodes`, which (per the [docs](https://facebook.github.io/lexical-ios/documentation/lexical/baseselection/insertnodes(nodes:selectstart:)/)):

> Attempts to intelligently insert an arbitrary list of Lexical nodes into the EditorState at the current Selection according to a set of heuristics that determine how surrounding nodes should be changed, replaced, or moved to accomodate the incoming ones.

I don't know exactly how to repro this (how to get `$getSelection()` to return `null`), but I've seen it happen and confirmed it in a debugger.