openctx: use the exact mention.uri for items
VSCode's URI.toString will escape the = and & from the query part of a
URL. You can optionally pass in skipEncode to toString, but in reality
we should be passing back the exact URL the provider supplied and should
not be potentially changing it at all.

This commit fixes the sourcegraph-search provider. Note that the links
displayed in the @-mention chip still will not work due to the above
escaping. I am unsure of a clean fix for that yet.

Test Plan: manually tested sourcegraph-search provider