Update recording for previously non-determinic enterprise test
Turns out that the Enterprise test case for the "Document" command was
non-determinic due to tree-sitter loading, and it should be determinstic
from now on. Turns out that we didn't correctly trim streaming replies
in the enterprise recording so this commit also shaves off a ton of
noise in the recording.