do not open enhanced context popover by default on new installs
This helped with discoverability but I'm removing this for the prerelease build because it seems like we'll change this UI a lot before the next release and it was causing flakiness. If we don't make progress on that new UI as hoped, then we probably need to add it back to avoid losing a lot of users who don't know how to set up context.

Added https://linear.app/sourcegraph/issue/CODY-1781/check-whether-we-need-to-go-back-to-showing-enhanced-context-popover to track this to make sure we don't miss it for the release.