message editor toolbar improvements to send buttons and auto code context
Often you want to be able to chat with no context or only explicitly @-mentioned context. This is helpful if, for example, you are asking about a general code concept or want to make extra sure that Cody only focuses on certain files when answering. This was previously possible with the hidden <kbd>Alt/Opt+Enter</kbd> shortcut, but that was hard to discover.

- Removes the checkbox from the "automatic code context" (fka "enhanced context settings") popover.
- There are now 2 send buttons, one (default) to send with automatic code context and one to send "w/o context".
- Makes all toolbar buttons tabbable for keyboard accessibility.