allow copy-to-clipboard of chat response while still streaming
Previously, if you tried to select text in Cody's response when it was still in progress (streaming in), it was not possible. Your selection would be immediately deselected.

Now, your selection is preserved and you can copy in-progress responses.

The icons below the code blocks do not yet show up until the message is complete. That will be fixed in a later commit.

This change was more involved than I expected. I also needed to:

- Switch to using a Markdown renderer that didn't just call `.innerHTML = ...`, so that we have stable DOM nodes that we can anchor our persistent selection to.
- Switch to a different way of sanitizing the HTML.
- Switch to a different way of syntax-highlighting.

Our tests for HTML sanitization still pass, which is a good sign of confidence.