Agent: bump commit (#1627)
This bumps Cody to a commit that includes this PR here
https://github.com/sourcegraph/cody/pull/4279

This PR fixes a tricky issue related to how
`vscode.TextEditor.selection` works. Previously, the text editor
selection was a constant, it never updated to reflect the latest
selection of the document. This could have caused buggy behavior in
scenarios like this:

```ts
for (const editor of vscode.window.visibleTextEditors) {
  // editor.selection is correct
  await parseDocument(editor.document)
  // editor.selection is outdated because we received a
  // `textDocument/didChange` in the meantime
}
```

## Test plan

n/a
<!-- All pull requests REQUIRE a test plan:
https://sourcegraph.com/docs/dev/background-information/testing_principles

Why does it matter?

These test plans are there to demonstrate that are following industry
standards which are important or critical for our customers.
They might be read by customers or an auditor. There are meant be simple
and easy to read. Simply explain what you did to ensure
your changes are correct!

Here are a non exhaustive list of test plan examples to help you:

- Making changes on a given feature or component:
- "Covered by existing tests" or "CI" for the shortest possible plan if
there is zero ambiguity
  - "Added new tests"
- "Manually tested" (if non trivial, share some output, logs, or
screenshot)
- Updating docs:
  - "previewed locally"
  - share a screenshot if you want to be thorough
- Updating deps, that would typically fail immediately in CI if
incorrect
  - "CI"
  - "locally tested"
-->