Testing: make the agent start during integration tests (#1628)
Previously, the agent wasn't starting in our tests because we didn't set
the right system property. Now, the agent manages to start correctly
during the tests.

While debugging this, I made a few other small improvements:

- Fix .tool-version file so we use the right node/pnpm version. Note
that this file had the wrong name (.tool-version instead of
.tool-versions)
- Move "Check updates" task off the blocking path during activation.
Starting the agent was taking ~2s longer than it should have because we
were checking if the Cody plugin had an available update before
starting. If you have a very slow network connection, this overhead
would have been even worse.

## Test plan

Tested locally. Confirmed that the run 2s faster.
<!-- All pull requests REQUIRE a test plan:
https://sourcegraph.com/docs/dev/background-information/testing_principles

Why does it matter?

These test plans are there to demonstrate that are following industry
standards which are important or critical for our customers.
They might be read by customers or an auditor. There are meant be simple
and easy to read. Simply explain what you did to ensure
your changes are correct!

Here are a non exhaustive list of test plan examples to help you:

- Making changes on a given feature or component:
- "Covered by existing tests" or "CI" for the shortest possible plan if
there is zero ambiguity
  - "Added new tests"
- "Manually tested" (if non trivial, share some output, logs, or
screenshot)
- Updating docs:
  - "previewed locally"
  - share a screenshot if you want to be thorough
- Updating deps, that would typically fail immediately in CI if
incorrect
  - "CI"
  - "locally tested"
-->