Minor change to indent code lenses to the line below them (#1604)
@danielmarquespt expressed that he would prefer to have the lenses
indent to the same level as the next non-blank line below them, rather
than always at a fixed offset.

This was very easy to add, and low-risk, so I added it. It handles
errors and falls back to the old indentation if there are any problems
at all.

## Test plan

I updated the test plan to include a visual inspection for the lens
indentation.