Agent: soften panic assertions
Previously, we would panic the agent process if the "source of truth"
document didn't match 100% between the client/server. Now, we soften the
assertion to only compare "source of truth" properties that are defined
by the client. This gives the client freedom to leave out some
properties (like content).

The motivation for this change is that the JB client can't always give a
100% correct answer in some scenarios. For example, on caret changes, we
only know the "logical position" (line/character), we don't know the
offset, or even the document contents. Just as we are calling
`Document.getText()`, the caret may already become outdated by mutation
on another thread.