Panic mode: bug fixes (#1640)
This commit addresses three issues with document synchronization:

- Panic mode now only sends properties that are guaranteed to be
correct. When paired with the changes in
https://github.com/sourcegraph/cody/pull/4291, this eliminates
false-positive panics. Previously, the submitted information and "source
of truth" information could already by out-of-sync before it even left
the JetBrains client due to the multithreaded nature of IDEA.
- Selection change handler now computes the selection from
`SelectionEvent` instead of `Editor` to eliminate the risk that the
editor's selection model changes in another thread as we're constructing
parameters for textDocument/didChange.
- File open handlers now use an `Editor` with the same URI as the
provided `VirtualFile`. Previously, we used `selectedTextEditor` that
could have a different URI.

## Test plan

Manually tested locally. I can still reproduce one panic with the
following minimized changes


https://github.com/sourcegraph/jetbrains/assets/1408093/18ceff9d-b919-4cd4-8812-88538c5e5461



<!-- All pull requests REQUIRE a test plan:
https://sourcegraph.com/docs/dev/background-information/testing_principles

Why does it matter?

These test plans are there to demonstrate that are following industry
standards which are important or critical for our customers.
They might be read by customers or an auditor. There are meant be simple
and easy to read. Simply explain what you did to ensure
your changes are correct!

Here are a non exhaustive list of test plan examples to help you:

- Making changes on a given feature or component:
- "Covered by existing tests" or "CI" for the shortest possible plan if
there is zero ambiguity
  - "Added new tests"
- "Manually tested" (if non trivial, share some output, logs, or
screenshot)
- Updating docs:
  - "previewed locally"
  - share a screenshot if you want to be thorough
- Updating deps, that would typically fail immediately in CI if
incorrect
  - "CI"
  - "locally tested"
-->

Co-authored-by: Piotr Kukielka <pkukielka@virtuslab.com>