feat(agent): add mode parameter to editCommands/code
This change adds a new `mode` parameter to the `editCommands/code` request in the agent protocol. The `mode` parameter can be either `'edit'` or `'insert'`, and it allows the client to specify whether the edit command should perform an edit or an insertion operation.

The corresponding changes have been made in the Kotlin binding and the TypeScript agent implementation to support this new parameter.