feat(vscode): add checks for empty selection in doc and test commands
This change adds checks for empty selections in the `executeDocCommand` and `executeTestEditCommand` functions. If the selected text is empty, the functions will throw an error to prevent the commands from executing on an empty selection.

The changes also include some minor logging improvements to provide more verbose information about the arguments being passed to the commands.