Agent: implement `vscode.TextDocument` with CodeMirror
While investigating JetBrains document synchronization issues, it became
obvious that we need a more robust and higher performance solution to
deal with incremental document changes on the agent server side.
Previously, the solution was a naive brute-force algorithm that
allocated a ton of strings for every single character edit. Now, we use
CodeMirror to handle all the editor state, including interpretation of
incremental document edits.

On a basic benchmark simulating editing a 50k loc file, the new solution
is ???x faster. Before this change, it took ???ms to make an edit, and
now it takes ???ms.