Fix: (edit) start DocumentCodeSession on Doc retry instead of EditCodeSession (#1649)
fix: add retry functionality to EditRetryAction

## Changes

The changes below fix the issue where clicking retry and edit after the
Document command would replace the current selection instead of
inserting.

- This change adds a new `mode` parameter to the `InlineEditParams` and
`EditCodeSession` classes. The `mode` parameter is used to indicate
whether the edit session is for an "insert" or "edit" operation.
- The `EditCommandPrompt` class has been updated to pass the appropriate
mode to the `EditCodeSession` based on whether the active session has
already inserted text.

This change allows us to start `EditCodeSession` in `insert` mode,
allowing users to edit the prompt used for the Doc command and then
insert the generated code to the file instead of replacing it in `edit`
mode.

## Test plan

<!-- All pull requests REQUIRE a test plan:
https://sourcegraph.com/docs/dev/background-information/testing_principles

Why does it matter?

These test plans are there to demonstrate that are following industry
standards which are important or critical for our customers.
They might be read by customers or an auditor. There are meant be simple
and easy to read. Simply explain what you did to ensure
your changes are correct!

Here are a non exhaustive list of test plan examples to help you:

- Making changes on a given feature or component:
- "Covered by existing tests" or "CI" for the shortest possible plan if
there is zero ambiguity
  - "Added new tests"
- "Manually tested" (if non trivial, share some output, logs, or
screenshot)
- Updating docs:
  - "previewed locally"
  - share a screenshot if you want to be thorough
- Updating deps, that would typically fail immediately in CI if
incorrect
  - "CI"
  - "locally tested"
-->

1. Select a function in yourfile
2. Execute the Document Code command 
3. Verify the `Edit & Retry` shows up
4. Click `Edit & Retry` to update prompt
5. Verify it behaves the same as if you first run the Document Code
command where the code would be insert instead of replacing your code
6. Click `Retry` a few more times to confirm
7. Click `Accept`
8. Execute `Edit Code` to edit the function
9. Verify it works as replace


## Demo



https://github.com/sourcegraph/jetbrains/assets/68532117/c778fdfe-874b-4847-8166-61fd2a5f2b09